Update overview.md
It is no longer accurate the at.js library is faster or better than the AEP web SDK for Target, so I changed the wording to show that. There is no problem in rephrasing what I wrote if you feel there is a better way to describe it. 
Also, https://experienceleague.adobe.com/docs/target-dev/developer/client-side/overview.html?lang=en - this page includes the same snippets and the modal won't pop up for me to suggest the changes there 
Thank you!